requirements.txt
python-telegram-bot==20.7
supabase==2.3.4
python-dotenv==1.0.1
httpx==0.26.0
